Separate shader-specific code from atom-specific code
This will help make it easy to run shaders in the browser.
