+++ lib/glsl-preview-view.js
@@ -4,40 +4,11 @@ const {clipboard, nativeImage} = require("electron");
 const {dialog} = require("remote");
 const fs = require("fs");
 
-const THREE = require("three");
-const {debounce, difference} = require("underscore");
+const {debounce} = require("underscore");
 const interact = require("interactjs");
 
 const Config = require("./config");
-
-function parseLineNumberFromErrorMsg(msg) {
-  const match = /ERROR: \d+:(\d+)/.exec(msg);
-  let lineNumber = null;
-  if (match && match[1]) {
-    lineNumber = parseInt(match[1], 10);
-  }
-  if (lineNumber !== null) {
-    const prologueLines = 107; // lines added before the user's shader code, by us or by THREE
-    return lineNumber - prologueLines;
-  }
-  return null;
-}
-
-function parseTextureDirectives(source) {
-  // Looking for lines of the form:
-  // uniform sampler2D foo; // ../textures/bar.jpg
-  const test = /^uniform sampler2D (\S+);\s*\/\/\s*(.+)$/gm;
-  const textureDirectives = [];
-  let match = test.exec(source);
-  while (match !== null) {
-    textureDirectives.push({
-      textureId: match[1],
-      filePath: match[2],
-    });
-    match = test.exec(source);
-  }
-  return textureDirectives;
-}
+const Shader = require("./shader");
 
 function getActiveTab() {
   return document.querySelector(".tab[data-type=\"GlslPreviewView\"]");
@@ -54,29 +25,10 @@ function getFileURI(editorPath, filePath) {
   return `file://${absPath}`;
 }
 
-function devicePixelRatio() {
-  return window.devicePixelRatio || 1;
-}
-
 function elementSize(elem) {
   const rect = elem.getBoundingClientRect();
   return [rect.width, rect.height];
 }
-
-const vertexShader = `
-  void main() {
-    gl_Position = vec4(position, 1.0);
-  }
-`;
-
-const defaultUniforms = `
-  uniform vec2 iResolution;
-  uniform vec2 iMouse;
-  uniform float iGlobalTime;
-  uniform vec2 u_resolution;
-  uniform vec2 u_mouse;
-  uniform float u_time;
-`;
 
 module.exports = class GlslPreviewView {
   constructor({editorId}) {
@@ -91,21 +43,38 @@ module.exports = class GlslPreviewView {
 
     this.emitter = new Emitter();
     this.disposables = new CompositeDisposable();
-    this.loaded = false;
 
     this.shaderError = null;
     this.textureError = null;
 
     this.lastSavePath = null;
 
-    // Setup webgl
-    this.renderer = new THREE.WebGLRenderer();
-    this.renderer.setPixelRatio(devicePixelRatio());
-
-    this.element.appendChild(this.renderer.domElement);
+    this.shader = new Shader({
+      buildTextureURL: filePath => getFileURI(this.getPath(), filePath),
+    });
+    this.shader.onShaderLoad = this.hideError.bind(this);
+    this.shader.onShaderError = this.showError.bind(this);
+    this.shader.onTextureLoad = () => {
+      // TODO: make errors not dismissable, relying on the error indicator in the editor
+      if (this.textureError) {
+        this.textureError.dismiss();
+      }
+    };
+    this.shader.onTextureError = (textureURL) => {
+      // TODO: set error on the line number of the texture
+      if (Config.showErrorMessage()) {
+        if (this.textureError) {
+          this.textureError.dismiss();
+        }
+        const opts = {detail: textureURL, dismissable: true};
+        this.textureError = atom.notifications.addError("Error loading texture", opts);
+      }
+    };
+    this.element.appendChild(this.shader.domElement);
+    this.disposables.add(this.shader);
 
     this.userSize = {};
-    interact(this.renderer.domElement).resizable({
+    interact(this.shader.domElement).resizable({
       edges: {left: true, right: true, bottom: true, top: true},
     }).on("resizemove", (event) => {
       const {width, height} = event.rect;
@@ -113,68 +82,7 @@ module.exports = class GlslPreviewView {
       this._onResize();
     });
 
-    this.scene = new THREE.Scene();
-
-    this.camera = new THREE.OrthographicCamera(-1, 1, 1, -1, 0.1, 10);
-    this.camera.position.z = 1;
-
-    this.renderer.render(this.scene, this.camera);
-
-    this.clock = new THREE.Clock(true);
-
-    this.uniforms = {
-      iGlobalTime: {value: 0},
-      iResolution: {value: new THREE.Vector2()},
-      iMouse: {value: new THREE.Vector2()},
-      u_time: {value: 0},
-      u_resolution: {value: new THREE.Vector2()},
-      u_mouse: {value: new THREE.Vector2()},
-    };
-
-    this.textures = [];
-
-    this.mesh = null;
-    this.geometry = new THREE.PlaneBufferGeometry(2, 2);
-
-    this._update = this._update.bind(this);
-    this._update();
-
-    this._onMouseMove = this._onMouseMove.bind(this);
     this._onResize = this._onResize.bind(this);
-    this.swapMesh = debounce(this.swapMesh.bind(this), 250);
-  }
-
-  swapMesh() {
-    const material = new THREE.ShaderMaterial({
-      uniforms: this.uniforms,
-      vertexShader,
-      fragmentShader: this.fragShader,
-    });
-
-    const tmpMesh = new THREE.Mesh(this.geometry, material);
-
-    this.scene.add(tmpMesh);
-
-    setTimeout(() => {
-      let diagnostics;
-      if (tmpMesh.material.program) {
-        diagnostics = tmpMesh.material.program.diagnostics;
-      }
-      if (diagnostics && !diagnostics.runnable) {
-        const msg = diagnostics.fragmentShader.log;
-        this.showError(msg, parseLineNumberFromErrorMsg(msg));
-        tmpMesh.material.dispose();
-        this.scene.remove(tmpMesh);
-      } else {
-        this.hideError();
-        if (this.mesh) {
-          // console.log("removing old mesh")
-          this.mesh.material.dispose();
-          this.scene.remove(this.mesh);
-        }
-        this.mesh = tmpMesh;
-      }
-    }, 100);
   }
 
   _onResize() {
@@ -191,34 +99,7 @@ module.exports = class GlslPreviewView {
       }
     }
 
-    const dpr = devicePixelRatio();
-
-    this.uniforms.iResolution.value.x = width * dpr;
-    this.uniforms.iResolution.value.y = height * dpr;
-    this.uniforms.u_resolution.value.x = this.uniforms.iResolution.value.x;
-    this.uniforms.u_resolution.value.y = this.uniforms.iResolution.value.y;
-
-    this.renderer.setSize(width, height);
-  }
-
-  _onMouseMove() {
-    const [width, height] = elementSize(this.renderer.domElement);
-
-    this.uniforms.iMouse.value.x = event.offsetX / width;
-    this.uniforms.iMouse.value.y = 1 - (event.offsetY / height);
-    this.uniforms.u_mouse.value.x = this.uniforms.iMouse.value.x;
-    this.uniforms.u_mouse.value.y = this.uniforms.iMouse.value.y;
-  }
-
-  _update() {
-    if (this.IS_DESTROYED) { return; }
-
-    requestAnimationFrame(this._update);
-
-    this.uniforms.iGlobalTime.value = this.clock.getElapsedTime();
-    this.uniforms.u_time.value = this.uniforms.iGlobalTime.value;
-
-    this.renderer.render(this.scene, this.camera);
+    this.shader.setSize(width, height);
   }
 
   attached() {
@@ -236,21 +117,8 @@ module.exports = class GlslPreviewView {
     if (this.IS_DESTROYED) { return; }
     this.IS_DESTROYED = true;
 
-    cancelAnimationFrame(this._update);
-
-    // remove all children
-    this.scene.children.forEach((child) => {
-      child.parent.remove(child);
-    });
-
-    this.renderer.domElement = null;
-
-    this.renderer = null;
-    this.scene = null;
-
-    this.element.innerHTML = "";
-
     this.disposables.dispose();
+    this.element.innerHTML = "";
   }
 
   resolveEditor(editorId) {
@@ -279,9 +147,6 @@ module.exports = class GlslPreviewView {
   }
 
   subscribeToEvents() {
-    this.renderer.domElement.addEventListener("mousemove", this._onMouseMove, false);
-    // Don't need to remove this, because we'll just remove the element.
-
     // FIXME: this is only fired for resizing the whole window, not just the panel.
     window.addEventListener("resize", this._onResize, false);
     this.disposables.add(new Disposable(() => {
@@ -323,10 +188,6 @@ module.exports = class GlslPreviewView {
 
   renderView() {
     if (this.IS_DESTROYED) { return; }
-
-    if (!this.loaded) {
-      this.showLoading();
-    }
     this.getShaderSource().then(this.updateShader.bind(this));
   }
 
@@ -338,15 +199,7 @@ module.exports = class GlslPreviewView {
   }
 
   updateShader(source) {
-    if (source !== null && source.length > 0) {
-      const parsedTextures = parseTextureDirectives(source);
-      const oldTextures = difference(this.textures, parsedTextures);
-      const newTextures = difference(parsedTextures, this.textures);
-      oldTextures.forEach(texture => this.removeTexture(texture.textureId));
-      newTextures.forEach(texture => this.addTexture(texture.filePath, texture.textureId));
-      this.fragShader = defaultUniforms + source;
-    }
-    this.swapMesh();
+    this.shader.updateShader(source);
   }
 
   onDidChangeTitle(callback) {
@@ -416,53 +269,9 @@ module.exports = class GlslPreviewView {
     }
   }
 
-  showLoading() {
-    this.loading = true;
-  }
-
-  addTexture(filePath, textureId) {
-    const fileURI = getFileURI(this.getPath(), filePath);
-
-    const onLoad = () => {
-      if (this.textureError) {
-        this.textureError.dismiss();
-      }
-      setTimeout(this.swapMesh, 100);
-    };
-
-    const onError = () => {
-      if (Config.showErrorMessage()) {
-        if (this.textureError) {
-          this.textureError.dismiss();
-        }
-        const opts = {detail: fileURI, dismissable: true};
-        this.textureError = atom.notifications.addError("Error loading texture", opts);
-      }
-    };
-
-    const texture = new THREE.TextureLoader().load(fileURI, onLoad, null, onError);
-    this.uniforms[textureId] = {value: texture};
-    this.textures.push({filePath, textureId});
-  }
-
-  removeTexture(textureId) {
-    const index = this.textures.findIndex(tex => tex.textureId === textureId);
-    if (index === -1) {
-      throw new Error("tried to remove a texture that doesn't exist");
-    }
-    this.textures.splice(index, 1);
-
-    this.uniforms[textureId].value.dispose();
-    this.uniforms[textureId].value.needsUpdate = true;
-
-    delete this.uniforms[textureId];
-
-    this.mesh.material.needsUpdate = true;
-  }
-
   getImage() {
-    this.renderer.render(this.scene, this.camera);
-    const dataURL = this.renderer.domElement.toDataURL();
+    this.shader.render();
+    const dataURL = this.shader.domElement.toDataURL();
     return nativeImage.createFromDataURL(dataURL);
   }
